feat: made script into a django custom command
